requirements.txt
Django
PIL
Pillow
South
django-ckeditor
django-filebrowser
django-grappelli
psycopg2
